Fixing issues

+++ Engine_Master_UPC/DebugDrawPass.cpp
@@ -466,7 +466,7 @@ private:
     ComPtr<ID3D12CommandAllocator>    uploadCommandAllocator;
     ComPtr<ID3D12Fence1>              uploadFence;
     HANDLE                            uploadEvent = NULL;
-    uint32_t                          uploadFenceValue = 0;
+    uint64_t                          uploadFenceValue = 0;
     D3D12_CPU_DESCRIPTOR_HANDLE       cpuTextHandle;
     D3D12_GPU_DESCRIPTOR_HANDLE       gpuTextHandle;
